requirements.txt
# for local llm
torch>=2.3.0
transformers
accelerate
huggingface_hub
sentencepiece
# for workflow
openai
google-genai
numpy
pandas
tqdm
pyyaml
loguru
